Devide test db in e2e tests
