+++ Makefile
@@ -45,7 +45,7 @@ docs: docs/conf.py $(RST_FILES)
 .PHONY: fmt
 fmt:
 	npm run fmt
-	black ./optuna_dashboard/ ./python_tests/
+	black ./optuna_dashboard/ ./python_tests/ ./e2e_tests/
 	isort .
 
 .PHONY: clean
